requirements.txt
pandas==1.5.1
streamlit
pickle
numpy
Pillow
fiddler-client==1.4.4
matplotlib
torch==1.8.0
